cstore: fixing implicit type conversion for building on Mac
On Mac, this currently fails with
cstore/uniondatapackstore.cpp:62:78: error: implicit conversion loses integer
      precision: 'data_offset_t' (aka 'unsigned long long') to 'size_t'
      (aka 'unsigned long') [-Werror,-Wshorten-64-to-32]
  ...ConstantStringRef((const char*)fulltextLink->delta, fulltextLink->delta_sz);
     ~~~~~~~~~~~~~~~~~                                   ~~~~~~~~~~~~~~^~~~~~~~

+++ cstore/uniondatapackstore.cpp
@@ -59,7 +59,8 @@ ConstantStringRef UnionDatapackStore::get(const Key &key) {
 
   // Short circuit and just return the full text if it's one long
   if (links.size() == 0) {
-    return ConstantStringRef((const char*)fulltextLink->delta, fulltextLink->delta_sz);
+    return ConstantStringRef((const char*)fulltextLink->delta,
+                             (size_t)fulltextLink->delta_sz);
   }
 
   std::reverse(links.begin(), links.end());
